Created new Model for functionality
